changed some ID's from int to long.

+++ trunk/src/main/java/com/beimin/eveapi/ApiAuth.java
@@ -7,7 +7,7 @@ public abstract class ApiAuth {
 
 	public abstract int getUserID();
 
-	public abstract int getCharacterID();
+	public abstract long getCharacterID();
 
 	public abstract String getApiKey();
